backend/pyproject.toml
[project]
name = "union-noise-backend"
version = "0.1.0"
description = "AI-assisted noise barrier design system — FastAPI backend"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "litellm>=1.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "google-genai>=1.0.0",
    "playwright>=1.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
